complete refactor + comments
