+++ gonk-database/src/strsim.rs
@@ -1,3 +1,4 @@
+//! Ripped from https://github.com/dguo/strsim-rs
 use std::cmp::{max, min};
 
 pub fn jaro_winkler(a: &str, b: &str) -> f64 {
